% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace.test}
\alias{ace.test}
\title{Short one-phrase description.}
\usage{
ace.test(snps, phen, tree, method = "discrete", snps.ace = FALSE)
}
\arguments{
\item{snps}{A matrix containing binary snps for all individuals.}

\item{phen}{A factor containing the phenotype for which to test for association.}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between individuals for which snps and phen are known.}

\item{method}{A character string specifying the type of ACE method to implement.}

\item{snps.ace}{A logical indicating whether to run ACE on all snps (TRUE, slower)
or to run parsimony on snps instead (FALSE, faster; the default).}
}
\description{
Longer proper discription of function...
}
\examples{

## example: ##

# load data
data("snps.ace")
data("phen.ace")
data("tree.ace")
snps <- snps.ace
phen <- phen.ace
tree <- tree.ace

## run ace.test on subset of data (otherwise SLOW!)
out <- ace.test(snps[,1:10], phen, tree, method="discrete")

## examine output
round(out, 4)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

