% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{asr}
\alias{asr}
\title{Short one-phrase description.}
\usage{
asr(var, tree, type = c("parsimony", "ace"), method = "discrete")
}
\arguments{
\item{var}{Either a matrix or a vector containing the state of a variable (eg. SNPs or a phenotype)
for all individuals (ie. for all terminal nodes in the tree).}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between the individuals for which snps and phen are known.}

\item{type}{A character string specifying whether ancestral state reconstruction should be
performed by \code{parsimony} or \code{ace} (as performed in package \emph{ape}).}

\item{method}{A character string specifying the type of ACE method to implement (only used if
\code{type} is set to "ace").}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

