% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{assoc.test}
\alias{assoc.test}
\title{Run a test of association between SNPs and a phenotype.}
\usage{
assoc.test(snps, phen, test = c("score", "cor", "fisher", "ace.cum",
  "ace.pagel"))
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{test}{A character string specifying the test to be used to measure correlations between the snps
(real and simulated) and the phenotype. Must be one of 'score', 'cor', 'fisher', 'ace.cum', 'ace.pagel'
(indicating, respectively, the correlation score, standard correlation, fisher's exact test,
ancestral character estimation (ACE) based cumulative multiplicative test, and ACE-based Pagel-inspired test).}
}
\description{
Run one of five tests of association between each column of a SNPs matrix and a phenotype
(some tests only implemented for \emph{binary} SNPs and phenotype).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

