% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeswarm.plot.R
\name{beeswarmPlot}
\alias{beeswarmPlot}
\title{Beeswarm-and-Box-Plot.}
\usage{
beeswarmPlot(y = "sensitivity", x = "test", df, y.lab = NULL,
  pt.size = 4, x.text = FALSE, x.text.size = 15, y.text.size = 15,
  y.title.size = 20, y.lim = c(-0.002, 1.02), mean = TRUE)
}
\arguments{
\item{y}{A character string specifying the label of the (numeric) column
in data frame \code{df} to be plotted along the y-axis.}
}
\description{
Wrapper combining the beeswarm and box plot functions from packages \code{beeswarm} and \code{ggplot2}.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
