% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{calculate.integrated.llk_scaled}
\alias{calculate.integrated.llk_scaled}
\title{Tree Analysis.
Calculate integrated likelihood
Scaled version}
\usage{
calculate.integrated.llk_scaled(jt.prior, llk.surf, scaled = TRUE)
}
\arguments{
\item{jt.prior}{input}

\item{llk.surf}{input}

\item{scaled}{input}
}
\value{
A numeric value.
}
\description{
Tree Analysis.
Calculate integrated likelihood
Scaled version
}
\examples{

}
