% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions_additive_model.R
\name{calculate.integrated.llk_scaled_additive}
\alias{calculate.integrated.llk_scaled_additive}
\title{calculate.integrated.llk_scaled_additive}
\usage{
calculate.integrated.llk_scaled_additive(prior, llk.surf, scaled = TRUE)
}
\arguments{
\item{prior}{input}

\item{llk.surf}{input}

\item{scaled}{input}
}
\description{
calculate.integrated.llk_scaled_additive
}
\examples{

}
