% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate.llk.grid.wrapper}
\alias{calculate.llk.grid.wrapper}
\title{Wrapper function to use with parallel package}
\usage{
calculate.llk.grid.wrapper(d)
}
\arguments{
\item{d}{List of objects needed by calculate_1d_llk_grid_scaled function}
}
\description{
Wrapper function to use with parallel package
}
\examples{


}
