% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{calculate.llk.grid_scaled}
\alias{calculate.llk.grid_scaled}
\title{Tree Analysis.
Calculate llk grid.
Scaled version}
\usage{
calculate.llk.grid_scaled(jt.prior, data, do.plot = FALSE, debug = FALSE,
  scaled = TRUE)
}
\arguments{
\item{jt.prior}{input}

\item{data}{input}

\item{do.plot}{input}

\item{debug}{input}

\item{scaled}{input}
}
\value{
A list.
}
\description{
Tree Analysis.
Calculate llk grid.
Scaled version
}
\examples{

}
