% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions_additive_model.R
\name{calculate.llk.grid_scaled_additive}
\alias{calculate.llk.grid_scaled_additive}
\title{Tree Analysis.
Calculate llk grid for the additive genetic model.
Scaled version}
\usage{
calculate.llk.grid_scaled_additive(prior, data, do.plot = FALSE,
  debug = FALSE, scaled = TRUE)
}
\arguments{
\item{prior}{input}

\item{data}{input}

\item{do.plot}{input}

\item{debug}{input}

\item{scaled}{input}
}
\value{
a list
}
\description{
Tree Analysis.
Calculate llk grid for the additive genetic model.
Scaled version
}
\examples{

}
