% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate_1d_llk_grid_scaled}
\alias{calculate_1d_llk_grid_scaled}
\title{Calculate LLK over the grid}
\usage{
calculate_1d_llk_grid_scaled(prior, x0, x1, scaled = TRUE)
}
\arguments{
\item{prior}{input}

\item{x0}{input}

\item{x1}{input}

\item{scaled}{input}
}
\description{
Calculate LLK over the grid
}
\examples{
calculate_1d_llk_grid_scaled() 

}

