% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate_1d_llk_grid_scaled_approx}
\alias{calculate_1d_llk_grid_scaled_approx}
\title{Calculate LLK over the grid with approximation}
\usage{
calculate_1d_llk_grid_scaled_approx(prior = NULL, x0 = NULL, x1 = NULL)
}
\arguments{
\item{prior}{1D prior}

\item{x0}{Values for unaffected individuals}

\item{x1}{Values for affected individuals}
}
\description{
Calculate LLK over the grid with approximation
}
\examples{

}
