% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate_1d_prior}
\alias{calculate_1d_prior}
\title{Function to generate a 1D non-local prior.
For use with GRS and additive genetic effects}
\usage{
calculate_1d_prior(b1.range = c(-4, 4), spac.b1 = 0.01, mn.b1 = 0,
  sd.b1 = 1, pi.1 = 0.001)
}
\arguments{
\item{b1.range}{input}

\item{spac.b1}{input}

\item{mn.b1}{input}

\item{sd.b1}{input}

\item{pi.1}{input}
}
\value{
A list with objects b.grid, prior and val.
}
\description{
Function to generate a 1D non-local prior.
For use with GRS and additive genetic effects
}
\examples{
calculate_1d_prior(b1.range=c(-4,4),spac.b1=0.01)

}
