% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{calculate_prior_2d}
\alias{calculate_prior_2d}
\title{Non-local prior for genetic effects. Scaled and reparametrised
Non-local prior of correlated effect sizes}
\usage{
calculate_prior_2d(b1.range = c(-10, 10), b2.range = c(-10, 10),
  spac.b1 = 0.1, spac.b2 = 0.1, mn.b1 = 0, mn.b2 = 0, sd.b1 = 2,
  sd.b2 = 4, cor = 0.5, k = 0.5, pi.1 = 0.001, f.off = 0.1,
  do.plot = FALSE)
}
\arguments{
\item{b1.range}{input b1 range}

\item{b2.range}{input b2 range}

\item{spac.b1}{input b1 grid density}

\item{spac.b2}{input b2 grid density}

\item{mn.b1}{input b1 mean}

\item{mn.b2}{input b2 mean}

\item{sd.b1}{input b1 sd}

\item{sd.b2}{input b2 sd}

\item{cor}{input correlation}

\item{k}{input k}

\item{pi.1}{input pi_1}

\item{f.off}{input TODO}

\item{do.plot}{input TODO}
}
\value{
A list
}
\description{
Non-local prior for genetic effects. Scaled and reparametrised
Non-local prior of correlated effect sizes
}
\examples{

}
