% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{cc_d.llk}
\alias{cc_d.llk}
\title{Function to calculate d(log likelihood)/db0
for logistic risk and full genetic model given b0, b1, b2, and data}
\usage{
cc_d.llk(b0 = b0, b1 = b1, b2 = b2, aff = aff, unaf = unaf)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{b2}{input}

\item{aff}{input}

\item{unaf}{input}
}
\value{
a nuemeric value
}
\description{
Function to calculate d(log likelihood)/db0
for logistic risk and full genetic model given b0, b1, b2, and data
}
\examples{

}
