% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{cc_llk}
\alias{cc_llk}
\title{Function to calculate log likelihood for logistic risk and full genetic
model given set of parameters.}
\usage{
cc_llk(b0 = b0, b1 = b1, b2 = b2, aff = aff, unaf = unaf)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{b2}{input}

\item{aff}{input}

\item{unaf}{input}
}
\value{
numeric value
}
\description{
Function to calculate log likelihood for logistic risk and full genetic
model given set of parameters.
}
\examples{

}
