% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions_additive_model.R
\name{cc_llk_additive}
\alias{cc_llk_additive}
\title{cc_llk_additive}
\usage{
cc_llk_additive(b0, b1, aff, unaf)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{aff}{input}

\item{unaf}{input}
}
\description{
cc_llk_additive
}
\examples{

}
