% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFML2treeWAS.R
\name{CFML2treeWAS.cont}
\alias{CFML2treeWAS.cont}
\title{(**Internal fn for running on cluster**)}
\usage{
CFML2treeWAS.cont(prefix)
}
\arguments{
\item{prefix}{A character string specifying the filename prefix of the dataset to be analysed.}
}
\description{
Run treeWAS for a continuous phenotype on processed CFML output. To run on the CLIMB cluster.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

