% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFML2treeWAS.R
\name{CFML2treeWAS}
\alias{CFML2treeWAS}
\title{(**Internal fn for running on cluster**)}
\usage{
CFML2treeWAS(prefix)
}
\arguments{
\item{prefix}{A character string specifying the filename prefix of the dataset to be analysed.}
}
\description{
Convert CFML output to treeWAS output with one function. To run on the CLIMB cluster.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

