% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent.sim.R
\name{coalescent.sim}
\alias{coalescent.sim}
\title{Short one-phrase description.}
\usage{
coalescent.sim(n.ind = 100, gen.size = 10000, sim.by = "locus",
  theta = 1 * 2, dist = NULL, theta_p = 15, phen = NULL,
  n.snp.assoc = 5, assoc.option = "all", assoc.prob = 90,
  haploid = TRUE, biallelic = TRUE, seed = NULL, plot = TRUE,
  heatmap = FALSE, plot2 = "UPGMA")
}
\arguments{
\item{n.ind}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

