% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{d.llk.d.b0_1d}
\alias{d.llk.d.b0_1d}
\title{Helper function}
\usage{
d.llk.d.b0_1d(b0, b1, x0, x1)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{x0}{input}

\item{x1}{input
d.llk.d.b0_1d()}
}
\description{
Helper function
}

