% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_0}
\alias{dist_0}
\title{Nsubs per site with no recombination (R = 0; M = 0.01).}
\format{A named vector of length 4.}
\usage{
data(dist_0)
}
\description{
This vector contains a homoplasy distribution,
representing the relative number of substitutions per site
that occurred along a phylogenetic tree when evolution was
simulated with a mutation rate of M = 0.01 and a recombination rate of R = 0
(\emph{r/m} = 0).
}
\details{
A per-site mutation rate of M = 0.01
indicates that each site, on average, undergoes 0.01
substitutions due to mutation along the phylogenetic tree.
A per-site recombination rate of R = 0
indicates that no recombination occurred.

Each element of the vector indicates the number of genetic loci
that have undergone the number of substitutions indicated by the name of that element (Nsub = i).

If visualised as a bar plot (with \code{barplot(dist_0)}),
one would see that the Nsub distribution is arranged as if it were the counts of a histogram
with index names along the x-axis, corresponding to Nsub (the number of substitutions per site),
and cell counts along the y-axis, showing Nloci (the number of genetic sites undergoing Nsub=i substitutions along the tree).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
