% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitch.sim.R
\name{fitch.sim}
\alias{fitch.sim}
\title{Short one-phrase description.}
\usage{
fitch.sim(snps, fitch.n.mts, tree, gen.size, biallelic = TRUE, seed = NULL)
}
\arguments{
\item{snps}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

