% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{fitMk}
\alias{fitMk}
\title{Short one-phrase description.}
\usage{
fitMk(tree, x, model = "SYM", fixedQ = NULL, ...)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

