% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{fitPagel}
\alias{fitPagel}
\title{Short one-phrase description.}
\usage{
fitPagel(tree, x, y, equal = FALSE, ...)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

