% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.snp.sim.R
\name{fwd.snp.sim}
\alias{fwd.snp.sim}
\title{Short one-phrase description.}
\usage{
fwd.snp.sim(n.snps = 10000, n.subs = 1, n.snps.assoc = 0,
  n.subs.assoc = 15, tree = coalescent.tree.sim(100),
  heatmap = FALSE, reconstruct = FALSE, dist.dna.model = "JC69",
  seed = 1)
}
\arguments{
\item{snps}{description.}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
