% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{get.ace.diffs}
\alias{get.ace.diffs}
\title{Short one-phrase description.}
\usage{
get.ace.diffs(var, edges)
}
\arguments{
\item{var}{A vector containing a variable whose change across edges we want to examine.}

\item{edges}{A 2-column matrix containing the upstream and downstream nodes
in columns 1 and 2 of a tree's edge matrix, as found in a phylo object's tree$edge slot.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
tree <- coalescent.tree.sim(n.ind = 100, seed = 1)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

