% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{get.ancestral.pars}
\alias{get.ancestral.pars}
\title{Ancestral sequence reconstruction via parsimony}
\usage{
get.ancestral.pars(var, tree)
}
\arguments{
\item{var}{A matrix or vector containing a variable whose state at ancestral nodes we want to infer.}

\item{tree}{A phylo object containing a phylogenetic tree whose tips contain the same individuals as are
in the elements of \code{var}, if \code{var} is a vector, or in the rows of \code{var}, if \code{var} is a matrix.}
}
\description{
A wrapper for the \code{ancestral.pars} function from \emph{ape}. Can perform
parsimonious ASR for variables in matrix or vector form.
}
\details{
Note that the (row)names of \code{var} should match the tip.labels of \code{tree}.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

