% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{get.assoc.scores}
\alias{get.assoc.scores}
\title{Get significant SNPs, according to a given test of association.}
\usage{
get.assoc.scores(
  snps,
  snps.sim,
  phen,
  tree,
  test = "terminal",
  correct.prop = FALSE,
  categorical = FALSE,
  snps.reconstruction = NULL,
  snps.sim.reconstruction = NULL,
  phen.reconstruction = NULL,
  unique.cols = FALSE
)
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{snps.sim}{A matrix or list of matrices containing simulated snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{tree}{A phylo object containing a phylogenetic tree in which the number of tips is equal to the
length of \code{phen} and the number of rows of \code{snps} and \code{snps.sim}.}

\item{test}{A character string or vector containing one or more of the following available tests of association:
"terminal", "simultaneous", "subsequent", "cor", "fisher". By default, the terminal test is run 
(note that within treeWAS, the first three tests are run in a loop by default). 
See details for more information on what these tests do and when they may be appropriate.}

\item{correct.prop}{A logical indicating whether the \code{"terminal"} and \code{"subsequent"} tests will be corrected for 
phenotypic class imbalance. Recommended if the proportion of individuals varies significantly across 
the levels of the phenotype (if binary) or if the phenotype is skewed (if continuous). 
If \code{correct.prop} is \code{FALSE} (the default), 
the original versions of each test will be run as described in our
\href{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005958}{PLOS Computational Biology paper}.
If \code{TRUE}, an alternate association metric (based on the phi correlation coefficient) is calculated 
across the terminal and all (internal and terminal) nodes, respectively.}

\item{categorical}{A logical indicating whether \code{phen} should be treated as a nominal categorical variable
whose unique values should be treated as levels rather than as meaningful numbers.}
}
\description{
Identify which SNPs are deemed to be significantly associated with a phenotype,
according to a given test of association and p-value.
(Serves as the treeWAS association testing function;
runs the \code{assoc.test} function internally.)
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
