% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitch.R
\name{get.fitch.n.mts}
\alias{get.fitch.n.mts}
\title{Caclulate parsimony scores.}
\usage{
get.fitch.n.mts(snps, tree)
}
\arguments{
\item{snps}{A numeric matrix containing two unique values with row.names matching tree tip.labels.}

\item{tree}{A phylo object.}
}
\description{
Determine parsimony scores for all genetic loci and a given tree.
An extension of the fitch function available in package phangorn.
}
\examples{

## generate a tree
tree <- rtree(100)
## geerate snps, a matrix of 0s and 1s
snps <- matrix(sample(c(0,1),100000,T), nrow=100)
row.names(snps) <- tree$tip.label

## run function
out <- get.fitch.n.mts(snps, tree)

## examine output
str(out)
table(out)
hist(out)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

