% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{get.n.subs}
\alias{get.n.subs}
\title{Short one-phrase description.}
\usage{
get.n.subs(n)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

