% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{get.posterior.node}
\alias{get.posterior.node}
\title{Tree Analysis.
Function to check B algorithm and
get credible sets}
\usage{
get.posterior.node(forward, backward, jt.prior, id = 1, plot = FALSE,
  return.ci = TRUE, verbose = FALSE, ci.level = 0.95, log.plot = TRUE)
}
\arguments{
\item{jt.prior}{input}

\item{llk.surf}{input}

\item{saled}{input}
}
\description{
Tree Analysis.
Function to check B algorithm and
get credible sets
}
\examples{
get.posterior.node()
}
