% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{get.posterior.node_1d}
\alias{get.posterior.node_1d}
\title{Estimate posterior probability and get MAP estimate}
\usage{
get.posterior.node_1d(forward, backward, prior, id = 1, plot = FALSE,
  return.ci = TRUE, verbose = FALSE, ci.level = 0.95, log.plot = TRUE)
}
\arguments{
\item{forward}{input}

\item{backward}{input}

\item{prior}{input}

\item{id}{input}

\item{plot}{input}

\item{return.ci}{input}

\item{verbose}{input}

\item{ci.level}{input}

\item{log.plot}{input}
}
\value{
A data.frame with the following columns: max_b, summed_llk, b_ci_lhs, b_ci_rhs, POST_ACTIVE
}
\description{
Estimate posterior probability and get MAP estimate
}
\examples{

}
