% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent.test.R
\name{get.score3}
\alias{get.score3}
\title{Short one-phrase description.}
\usage{
get.score3(Pa, Pd, Sa, Sd, l = NULL)
}
\arguments{
\item{Pa}{A numeric value containing either the state,
or the probability of the state, of the phenotype at a given \emph{ancestral} node.}

\item{Pd}{A numeric value containing either the state,
or the probability of the state, of the phenotype at a given \emph{descendant} node.}

\item{Sa}{A numeric value containing either the state,
or the probability of the state, of SNPi at a given \emph{ancestral} node.}

\item{Sd}{A numeric value containing either the state,
or the probability of the state, of SNPi at a given \emph{descendant} node.}

\item{l}{A numeric value specifying the length of the branch in the phylogenetic tree
that joins the ancestral and descendant node.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
tree <- coalescent.tree.sim(n.ind = 100, seed = 1)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

