% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{get.sig.snps}
\alias{get.sig.snps}
\title{Get significant SNPs, according to a given test of association.}
\usage{
get.sig.snps(snps, snps.sim, phen, tree, test = c("score", "cor", "fisher",
  "ace.cum", "ace.pagel"), p.value = 0.001, p.value.correct = c("bonf",
  "fdr", FALSE), p.value.by = c("count", "density"))
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{snps.sim}{A matrix or list of matrices containing simulated snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{tree}{A phylo object containing a phylogenetic tree in which the number of tips is equal to the
length of \code{phen} and the number of rows of \code{snps} and \code{snps.sim}.}

\item{test}{A character string specifying the test to be used to measure correlations between the snps
(real and simulated) and the phenotype. Must be one of 'score', 'cor', 'fisher' (indicating, respectively,
the correlation score, standard correlation, fisher's exact test).}

\item{p.value}{A single number specifying the p.value below which correlations are deemed to be 'significant'.}

\item{p.value.correct}{Specify if/how to correct for multiple testing:
either FALSE, or one of 'bonf' or 'fdr' (indicating, respectively,
the Bonferroni and False Discovery Rate corrections).}

\item{p.value.by}{Specify how to determine the location of the p.value threshold:
either 'count' or 'density' (indicating, respectively, that the p.value threshold should
be determined by exact count or with the use of a density function).}
}
\description{
Identify which SNPs are deemed to be significantly associated with a phenotype,
according to a given test of association and p-value.
(Serves as the treeWAS association testing function;
runs the \code{assoc.test} function internally.)
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn

fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

