% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot.bg}
\alias{ggplot.bg}
\title{Mimic ggplot2 Background}
\usage{
\method{ggplot}{bg}(bg = transp("lightgray", 0.5), x.ax = FALSE,
  y.ax = FALSE, box = TRUE, grid = TRUE, grid.col = "white",
  grid.nx = NULL, grid.ny = NULL, grid.lwd = 1, grid.lty = 1)
}
\description{
Get an imitation ggplot2-style background for plots made outside ggplot2
}
\details{
This function must be sandwiched between two instances
of the function used to generate the (foreground) plot
to which you are hoping to add this background.
\emph{Before} running the \code{ggplot.bg} function, you need to run your plot function
so that \code{ggplot.bg} knnows how to set the axes.
\emph{After} running the \code{ggplot.bg} function, you need to run your plot function
again \emph{with the added argument} \code{add=TRUE}
so that your plot can be overlayed on top of the background.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

