% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotbg}
\alias{ggplotbg}
\title{Mimic ggplot2 Background}
\usage{
ggplotbg(bg = transp("lightgray", 0.5), x.ax = FALSE, y.ax = FALSE,
  box = TRUE, grid = TRUE, grid.col = "white", grid.nx = NULL,
  grid.ny = NULL, grid.lwd = 1, grid.lty = 1)
}
\arguments{
\item{bg}{The background colour, by default ``lightgray'' with 50\% transparency.}

\item{x.ax}{A logical specifying whether to re-draw the x-axis.}

\item{y.ax}{A logical specifying whether to re-draw the y-axis.}

\item{box}{A logical specifying whether to draw a box around the plotting area.}

\item{grid}{A logical specifying whether to draw a grid across the background within the plotting area.}

\item{grid.col}{The color of the gridlines, ``white'' by default. Only used if grid is set to TRUE.}

\item{grid.nx}{An optional integer to specify the number of gridlines to be drawn along the x-axis.}

\item{grid.ny}{An optional integer to specify the number of gridlines to be drawn along the y-axis.}

\item{grid.lwd}{An integer specifying the lwd (line weight) of the gridlines; by default, set to 1.}

\item{grid.lty}{An integer specifying the line type to be used for the gridlines; by default, set to 1 (i.e., solid lines).}
}
\description{
Get an imitation ggplot2-style background for plots made outside ggplot2
}
\details{
This function must be sandwiched between two instances
of the function used to generate the (foreground) plot
to which you are hoping to add this background.
\emph{Before} running the \code{ggplot.bg} function, you need to run your plot function
so that \code{ggplot.bg} knnows how to set the axes.
\emph{After} running the \code{ggplot.bg} function, you need to run your plot function
again \emph{with the added argument} \code{add=TRUE}
so that your plot can be overlayed on top of the background.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
