% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.DNAbin.R
\name{heatmap.DNAbin}
\alias{heatmap.DNAbin}
\title{Short one-phrase description.}
\usage{
heatmap.DNAbin(dna, dist.dna.model = "JC69")
}
\arguments{
\item{dna}{A DNAbin object.}

\item{dist.dna.model}{A character string specifying the type of model to use in
calculating the genetic distance between individual genomes (see ?dist.dna).}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

