% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepFirstN}
\alias{keepFirstN}
\title{Truncate to keep only the \emph{first} N characters.}
\usage{
keepFirstN(x, n)
}
\arguments{
\item{x}{A vector whose element(s) will be truncated.}

\item{n}{An integer specifying the number of characters to \emph{keep}.}
}
\description{
Truncate an element, or each element of a vector, by
removing all but the first N characters of each element.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
