% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{llk_1d}
\alias{llk_1d}
\title{Likelihood function}
\usage{
llk_1d(b0, b1, x0, x1)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{x0}{input}

\item{x1}{input}
}
\description{
Likelihood function
}
\examples{

}
