% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{matrix.table}
\alias{matrix.table}
\title{Short one-phrase description.}
\usage{
matrix.table(data)
}
\arguments{
\item{data}{A matrix or data.frame.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
tree <- coalescent.tree.sim(n.ind = 100, seed = 1)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

