% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{memfree}
\alias{memfree}
\title{Get the current amount of available memory.}
\usage{
memfree(OS = NULL)
}
\arguments{
\item{OS}{A character string indicating the operating system of the machine in question.
Can be one of "Windows", "Mac" (or "Darwin"), or "Linux". If OS is NULL (the default),
OS will be set to Sys.info()["sysname"].}
}
\description{
Function to determine how much memory (in GB) is currently available for use
on your PC.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
