% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{mexp}
\alias{mexp}
\title{Short one-phrase description.}
\usage{
mexp(x, type = "spectral decomposition", t = 1, n = 20, k = 3)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

