% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFML2treeWAS.R
\name{PA.treeWAS}
\alias{PA.treeWAS}
\title{(**Internal fn for running on cluster**)}
\usage{
PA.treeWAS(prefix)
}
\arguments{
\item{prefix}{A character string specifying the filename prefix of the dataset to be analysed.}
}
\description{
Run treeWAS on Presence/Absence data. To run on the CLIMB cluster.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

