% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.tests.R
\name{pair.tests}
\alias{pair.tests}
\title{Pairwise tests for categorical phenotypes}
\usage{
pair.tests(x, y, z, method = "bonf", digits = 3)
}
\arguments{
\item{x}{A contingency table (snps[,i] x phen) for score 1 (\code{terminal.test} 
with \code{correct.prop = TRUE}, \code{categorical = TRUE}).}

\item{y}{A vector of values containing pairwise score 2 (\code{simultaneous.test} 
with \code{categorical = TRUE}) results for snps[,i].}

\item{z}{A contingency table (snps.rec[,i] x phen.rec) for score 3 (\code{subsequent.test} 
with \code{correct.prop = TRUE}, \code{categorical = TRUE}).}
}
\description{
Internal function to calculate treeWAS 
terminal, simultaneous, subsequent tests, 
and chi-squared p-values for a given snp across pairs of
phenotype levels.
}
\examples{
## Example ##
\dontrun{
## basic use of fn
out <- pair.tests(x, y, z)
}

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
