% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phen.cont.rank}
\alias{phen.cont.rank}
\title{A rank-transformed continuous phenotype.}
\format{
A named vector of length 533.
}
\usage{
data(phen.cont.rank)
}
\description{
This vector specifies the phenotype of each individual.
In this case, the phenotype is a rank, that has been derived by
rank-ordering the elements of the original continuous phenotype (\code{data(phen.cont)})
from lowest to highest.
Transforming by rank prior to analysis by treeWAS can be useful
for continuous phenotypic variables that are highly skewed or contain significant outliers
(see the treeWAS vignette).
% (see \code{vignette("treeWAS")}).
}
\details{
Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one element of the phenotypic vector.
Each element of the phenotypic vector gives the phenotypic value of the named individual.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
