% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phen.cont}
\alias{phen.cont}
\title{A continuous phenotype.}
\format{
A named vector of length 533.
}
\usage{
data(phen.cont)
}
\description{
This vector specifies the phenotype of each individual.
In this case, the phenotype is a continuous numeric value.
}
\details{
Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one element of the phenotypic vector.
Each element of the phenotypic vector gives the phenotypic value of the named individual.

Note that, due to some skew in the distribution of this continuous variable,
it may be useful to transform the phenotype by rank prior to analysis by treeWAS,
as in \code{data(phen.cont.rank)} (see the treeWAS vignette).
% (see \code{vignette("treeWAS")}).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
