% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phen}
\alias{phen}
\title{A binary phenotype.}
\format{
A named vector of length 100.
}
\usage{
data(phen)
}
\description{
This vector specifies the phenotype of each individual.
In this case, the phenotype is a binary variable.
Because the phenotypic vector is encoded as a factor
with two possible phenotypic states, "A" and "B",
which may be represented by the numeric values 1 and 2 (as in \code{str(phen)}).
}
\details{
Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one element of the phenotypic vector.
Each element of the phenotypic vector gives the phenotypic value of the named individual.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
