% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phen.reconstruction}
\alias{phen.reconstruction}
\title{The ancestral state reconstruction of a binary phenotype.}
\format{
A named vector of length 199.
}
\usage{
data(phen.reconstruction)
}
\description{
This vector contains the terminal and ancestral states of a binary phenotypic variable (\code{data(phen)}).
The observed phenotypic states of sampled individuals
(i.e., those represented at the terminal nodes of a phylogenetic tree)
are presented first, in elements 1:N (here 1:100).
The unobserved ancestral states of the phenotype at internal nodes have been
inferred via ancestral state reconstruction, using \code{asr(phen, tree)}.
}
\details{
Like the original phenotypic vector (\code{data(phen)}),
\code{phen.reconstruction} is a binary variable that is encoded as a factor
with two possible phenotypic states, "A" and "B",
which may be represented by the numeric values 1 and 2 (as in \code{str(phen.reconstruction)}).

Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one element of the phenotypic vector.
(Internal node names have been generated during ancestral state reconstruction.)
Each element of the phenotypic vector gives the phenotypic value of the named individual.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
