% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.plot.prob.phen.R
\name{plot.prob.phen}
\alias{plot.prob.phen}
\title{Plot the probability of association, given \code{p} and \code{n.snps.assoc}.}
\usage{
plot.prob.phen(p = 0.5, n.snps.assoc = 10)
}
\arguments{
\item{p}{A numeric value indicating the probability of substitution, at each site, along the tree.}

\item{n.snps.assoc}{An integer specifying the number of genetic loci that are associated with the phenotype.}
}
\description{
[*For use with the 'fwd.-.sim' functions:*] 
Plot the cumulative probability of association (Pr(phen=1)), with a given value of \code{p},
as the number of associated sites (SNPi=1) increases from i=0 to i=\code{n.snps.assoc}.
}
\examples{
\dontrun{
## basic use of fn ##
## compare probability of having phenotype with 10 SNPs at varying p:
plot.prob.phen(p=0.8, n.snps.assoc=10)
plot.prob.phen(p=0.5, n.snps.assoc=10)
plot.prob.phen(p=0.2, n.snps.assoc=10)
}
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
