% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sig.snps.R
\name{plot.sig.snps}
\alias{plot.sig.snps}
\title{Short one-phrase description.}
\usage{
plot.sig.snps(corr.dat, corr.sim, sig.corrs, sig.snps, sig.thresh, test,
  plot.null.dist = TRUE, plot.dist = FALSE)
}
\arguments{
\item{arg}{Description.}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

