% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{prepare_count_table}
\alias{prepare_count_table}
\title{Preparing count table from phenotype table}
\usage{
prepare_count_table(M, N, sparse_phe_M_by_N, ge)
}
\arguments{
\item{M}{number of disease}

\item{N}{number of individual}

\item{sparse_phe_M_by_N}{is a two column matrix, to summaries which disease
is carried by which patient. The first column indicate the index of the
diease, and the second column indicates patient index.}

\item{ge}{is a indicator vector length N, to identify if the nth patient
carries an alternative allele at the target position.}
}
\value{
An M by 6 matrix. Each column show the number of counts of genotype
and phenotype pairs of "0-0", "1-0", "2-0", "0-1", "1-1", "2-1"
}
\description{
Let M be the number of disease, and N be the number of individual. We can
construct a phenotype indicator matrix to show which person carries which
disease
}
\examples{
N = 1000
M = 500
N_by_M = N * M
phe = matrix(round(round(rexp(N_by_M)/5) > 0), nrow = N)
ge = matrix(round(runif(N) * 2), nrow = N)
sparse_phe = which(phe == 1, arr.ind=TRUE) # patient idx, phe idx
sparse_phe_M_by_N = sparse_phe[,c(2,1)] # phe idx, patient idx
count_table = prepare_count_table(M, N, sparse_phe_M_by_N, ge)

}
