% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{prepare_tree_table}
\alias{prepare_tree_table}
\title{Function to prepare Tree table for analysis}
\usage{
prepare_tree_table(phens = phens, tree = tree, samples = samples,
  NODE.COUNT.LIMIT = 10, NUM.CASES.MIN = 10)
}
\arguments{
\item{phens}{input}

\item{tree}{input}

\item{NODE.COUNT.LIMIT}{input}
}
\description{
Function to prepare Tree table for analysis
}
\examples{
prepare_tree_table()

}
