% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{prepare_tree_table_grs}
\alias{prepare_tree_table_grs}
\title{Function to prepare Tree table for analysis}
\usage{
prepare_tree_table_grs(phens = phens, tree = tree, NODE.COUNT.LIMIT = 10)
}
\arguments{
\item{phens}{input}

\item{tree}{input}

\item{NODE.COUNT.LIMIT}{input}
}
\value{
A data.frame with the tree data sorted for TreeWAS analysis
}
\description{
Function to prepare Tree table for analysis
}
\examples{

}
