% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{print.fitPagel}
\alias{print.fitPagel}
\title{Short one-phrase description.}
\usage{
\method{print}{fitPagel}(x, ...)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

