% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.R
\name{Q2P}
\alias{Q2P}
\title{Short one-phrase description.}
\usage{
Q2P(Q, type = NULL)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

