% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFML.R
\name{read.CFML}
\alias{read.CFML}
\title{Convert ClonalFrameML output.}
\usage{
read.CFML(prefix, tree = NULL, plot = TRUE)
}
\arguments{
\item{prefix}{A character string containing the prefix of all file names to be read in.}
}
\description{
Convert the output of ClonalFrameML into a form usable within \code{treeWAS}.
}
\details{
The \code{prefix} must be the prefix to three files ending in:
(i) "labelled_tree.newick", (ii) "ML_sequence.fasta", (iii) "position_cross_reference.txt".
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

