% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFML.R
\name{readCFML}
\alias{readCFML}
\title{Short one-phrase description.}
\usage{
readCFML(prefix)
}
\arguments{
\item{prefix}{A character string containing the prefix of all file names to be read in.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

