% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeFirstN}
\alias{removeFirstN}
\title{Truncate to remove all of the \emph{first} N characters.}
\usage{
removeFirstN(x, n)
}
\arguments{
\item{x}{A vector whose element(s) will be truncated.}

\item{n}{An integer specifying the number of characters to \emph{remove}.}
}
\description{
Truncate an element, or each element of a vector, by
removing the first N characters of each element.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
