% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectBiallelicSNP}
\alias{selectBiallelicSNP}
\title{Short one-phrase description.}
\usage{
selectBiallelicSNP(x, DNA = TRUE)
}
\arguments{
\item{x}{A character vector of length 1 containing a nucleotide to be converted.}

\item{DNA}{logical; if TRUE (default), uses DNA bases (ACGT), if FALSE, uses RNA bases (ACGU).}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

