% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.args}
\alias{set.args}
\title{Set a list of arguments.}
\usage{
set.args(args, envir = sys.frame(which = 0L))
}
\arguments{
\item{args}{A named list of arguments.}

\item{envir}{The environment in which these arguments will set.}
}
\description{
Function to set a list of arguments without having to remove commas.
Useful for troubleshooting. For example, if attempting to run a function
(particualrly one with many arguments) line by line,
\code{set.args} can be used to set a list of arguments in one go, by copying a
comma-separated set of arguments from an existing function call or a new call to \code{args(fn)}.
}
\details{
Please note that unless the \code{envir} argument is changed from its default (\code{sys.frame}),
any arguments set with \code{set.args} will \emph{over-ride} any values currently assigned to those names.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

