% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.test.R
\name{simultaneous.test}
\alias{simultaneous.test}
\title{Simultaneous test}
\usage{
simultaneous.test(
  snps.reconstruction,
  phen.reconstruction,
  tree,
  categorical = FALSE
)
}
\arguments{
\item{snps.reconstruction}{A matrix containing the terminal and reconstructed
ancestral states of SNPs for all nodes in the tree.}

\item{phen.reconstruction}{A vector containing the terminal and reconstructed
ancestral states of the phenotype for all nodes in the tree.}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between the individuals for which snps and phen are known.}
}
\description{
Calculates treeWAS score 2, the simultaneous test, as the number of 
substitutions or changes in genotype (\code{snps.reconstruction}) and phenotype 
(\code{phen.reconstruction}) that occur simultaneously on the same branches of the tree.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
