% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.sim.R, R/snp.sim_NEW.R, R/snp.sim_OLD.R
\name{snp.sim}
\alias{snp.sim}
\alias{snp.sim}
\alias{snp.sim}
\title{Short one-phrase description.}
\usage{
snp.sim(n.snps = 10000, n.subs = 1, snp.root = NULL, n.snps.assoc = 0,
  assoc.prob = 100, tree = coalescent.tree.sim(100), phen.loci = NULL,
  heatmap = FALSE, reconstruct = FALSE, dist.dna.model = "JC69",
  row.names = NULL, set = NULL, seed = 1)

snp.sim(n.snps = 10000, n.subs = 1, snp.root = NULL, n.snps.assoc = 0,
  assoc.prob = 100, tree = coalescent.tree.sim(100), phen.loci = NULL,
  heatmap = FALSE, reconstruct = FALSE, dist.dna.model = "JC69",
  row.names = NULL, set = NULL, seed = 1)

snp.sim(n.snps = 10000, n.subs = 1, snp.root = NULL, n.snps.assoc = 0,
  assoc.prob = 100, tree = coalescent.tree.sim(100), phen.loci = NULL,
  heatmap = FALSE, reconstruct = FALSE, dist.dna.model = "JC69",
  row.names = NULL, set = NULL, seed = 1)
}
\arguments{
\item{snps}{description.}

\item{snps}{description.}

\item{snps}{description.}
}
\description{
Longer proper discription of function...

Longer proper discription of function...

Longer proper discription of function...
}
\examples{
## Example ##

## Example ##

## Example ##

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}

Caitlin Collins \email{caitiecollins@gmail.com}

Caitlin Collins \email{caitiecollins@gmail.com}
}
