% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps.assoc}
\alias{snps.assoc}
\title{The phenotypically-associated sites in the \code{snps} matrix.}
\format{
A vector of length 10.
}
\usage{
data(snps.assoc)
}
\description{
This vector specifies the identities (names) of the loci in the genetic data matrix (see: \code{data(snps)})
that have been simulated along the phylogenetic tree (see: \code{data(tree)})
to be in statistical association with the phenotype (see: \code{data(phen)}).
Comparing this vector of snps column names to the set of snps loci identified by treeWAS
allows us to evaluate the performance of the treeWAS GWAS method.
After applying treeWAS to the components of this dataset,
using: \code{treeWAS(snps, phen, tree)},
we can assess the ability of treeWAS to recover these "known" associated sites
via any of its three association scores
(see the treeWAS vignette).
% (see vignette("treeWAS")).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
