% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps}
\alias{snps}
\title{A genetic data matrix.}
\format{A binary matrix with 100 rows and 20,003 columns.}
\usage{
data(snps)
}
\description{
This binary matrix contains the allelic states of genetic variables,
typically single-nucleotide polymorphisms (SNPs) (or the presence/absence states of accessory genes),
showing individuals in the rows and genetic loci in the columns.
}
\details{
Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one row of the snps matrix.
Each genetic locus is also required to have a unique name.

In this \code{snps} matrix, redundant columns are present for biallelic loci,
denoting the state of the second allele as the inverse of the previous column
(e.g., compare locus 1.g and locus 1.a).
These biallelic sites can be condensed into a more efficient binary form
by using \code{get.binary.snps(snps)}
(see the treeWAS vignette).
% (see vignette("treeWAS")).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
