% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps.reconstruction}
\alias{snps.reconstruction}
\title{The ancestral state reconstruction of a genetic data matrix.}
\format{
A binary matrix with 199 rows and 20,003 columns.
}
\usage{
data(snps.reconstruction)
}
\description{
This binary matrix contains the terminal and ancestral allelic states of a set of genetic variables
(for the original genetic data matrix, see: \code{data(snps)}),
showing individuals in the rows and genetic loci in the columns.
The observed genotypic states of sampled individuals
(i.e., those represented at the terminal nodes of a phylogenetic tree)
are presented first, in elements 1:N (here 1:100).
These rows of the matrix are identical to the input \code{snps} matrix (see: \code{data(snps)}).
The unobserved ancestral states of the genotype at internal nodes have been
inferred via ancestral state reconstruction, using \code{asr(snps, tree)}.
}
\details{
Each individual in the sample is represented by a unique identifier (name)
which corresponds to the name of one row of the snps matrix.
(Internal node names have been generated during ancestral state reconstruction.)
Each genetic locus is also required to have a unique name.

In this \code{snps.reconstruction} matrix, redundant columns are present for biallelic loci,
denoting the state of the second allele as the inverse of the previous column
(e.g., compare locus 1.g and locus 1.a).
These biallelic sites can be condensed into a more efficient binary form
by using \code{get.binary.snps(snps)}
(see the treeWAS vignette).
% (see vignette("treeWAS")).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
