% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.test.epi.R
\name{terminal.test.epi}
\alias{terminal.test.epi}
\title{Test for epistasis between genetic loci with Score 1.}
\usage{
terminal.test.epi(snps, snps.subset = NULL)
}
\arguments{
\item{snps}{A matrix containing the states of SNPs (in columns) for all individuals (in rows).}

\item{snps.subset}{An optional vector (see details); else, NULL. 
The snps.subset vector can be a character vector, containing a subset of colnames(snps.rec), 
a logical vector, using TRUE or FALSE to indicate which columns are to be retained and excluded,
or an integer vector, specifying the column indices to be retained.}
}
\description{
[*\emph{A work in progress; not curently integrated into treeWAS:}*]
Use the terminal.test (Score 1) to test for associations between genetic loci, 
which may indicate an epistatic interaction.
This function can be used either to test 
for pairwise association between all pairs of genetic loci
or for associations between a subset of snps and all other snps 
(recommended for large datasets; see details).
}
\details{
The number of pairwise tests between all pairs of snps 
grows rapidly as the number of snps columns increases. 
As such, for datasets where ncol(snps.reconstruction) is large, we recommend that
the snps.subset argument is used to reduce the number of tests, by
indicating which snps to test for association with all other snps. 
The snps.subset index can be used to select any subset of snps of interest. 
For example, one may wish to test for interactions between all snps and a subset of snps that 
had been deemed significantly associated with a particular phenotype in a previous run of treeWAS.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
