% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.test.R
\name{terminal.test}
\alias{terminal.test}
\title{Terminal test}
\usage{
terminal.test(snps, phen, correct.prop = FALSE, categorical = FALSE)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Calculates treeWAS score 1, the terminal test.
}
\examples{
## Example ##
\dontrun{
## basic use of fn
out <- terminal.test(snps, phen)
} 

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
