% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{A phylogenetic tree.}
\format{A phylo object with 100 terminal nodes and 99 internal nodes.}
\usage{
data(tree)
}
\description{
This phylogenetic tree is a phylo object (see \code{vignette("Trees", package="phangorn")})
connecting the individuals represented in
the rows of the example genetic data matrix (see: \code{data(snps)})
and the elements of the example phenotypic vector (see: \code{data(phen)}).
}
\details{
In this case, the tree was generated via simulation and used to simulate the genotypic and phenotypic data.
In a typical empirical analysis, however, a phylogenetic tree would represent the inferred
ancestral relationships between individuals, and it would be estimated from the available genetic data.
For example, such a phylogeny could be reconstructed using \code{tree.reconstruct(snps, method="NJ")},
or automatically generated within treeWAS, according to the \code{tree} argument, as in:
\code{treeWAS(snps, phen, tree="NJ")}
(see the treeWAS vignette).
% (see vignette("treeWAS")).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
