% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.sim.R
\name{tree.sim}
\alias{tree.sim}
\title{Short one-phrase description.}
\usage{
tree.sim(n.ind = length(tree$tip.label), tree = tree, gen.size = 10000,
  sim.by = "locus", theta = NULL, dist = NULL, theta_p = NULL,
  phen = NULL, biallelic = TRUE, seed = NULL, plot = TRUE,
  heatmap = FALSE, plot2 = "UPGMA")
}
\arguments{
\item{n.ind}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

