% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treeWAS.example.out}
\alias{treeWAS.example.out}
\title{Example output of treeWAS.}
\format{A treeWAS class object, comprising a list of length 5.}
\usage{
data(treeWAS.example.out)
}
\description{
This "treeWAS" class object is a list containing the output of a treeWAS analysis.
This GWAS analysis was performed to identify associations between
loci in the example genetic data matrix (see: \code{data(snps)})
and phenotypic states in the example phenotypic vector (see: \code{data(phen)}),
along the phylogenetic tree (see: \code{data(tree)}).
}
\details{
This \code{treeWAS} output was returned by the function:
\code{treeWAS(snps, phen, tree)}.
treeWAS output contains elements describing
the significant associations identified by each of the
three association scores applied to all genetic loci.
Additional elements of the output return all data that was
used in the GWAS analysis, including both data input to treeWAS
and all relevant data generated by treeWAS.

For a detailed description of the elements of this output,
please scroll down to the "Value" section of the \code{treeWAS} function documentation,
which can be accessed with: \code{?treeWAS}.
More information can also be found in the treeWAS vignette.
% (see vignette("treeWAS")).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
\keyword{data}
\keyword{datasets}
