% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.R
\name{write.treeWAS}
\alias{write.treeWAS}
\title{Write \code{treeWAS} output to a CSV file.}
\usage{
write.treeWAS(x, filename = "./treeWAS_results")
}
\arguments{
\item{x}{The output returned by \code{treeWAS}.}

\item{filename}{A character string containing the path and filename to which the .csv file will be saved;
by default, \code{filename = "./treeWAS_results"} and so would be saved to the current working directory.}
}
\description{
Save the results of \code{treeWAS} to a CSV file as a summary table of significant findings and scores
(excluding longer data elements within the output).
.
}
\examples{
## Example ##
\dontrun{
## Load data:
data(snps)
data(phen)
data(tree)

## Run treeWAS:
out <- treeWAS(snps, phen, tree, seed = 1)

## Save results to home directory:
write.treeWAS(x = out, filename = "~/treeWAS_results")
}

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
