% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace.test}
\alias{ace.test}
\title{Short one-phrase description.}
\usage{
ace.test(snps, phen, tree, method = "discrete")
}
\arguments{
\item{data}{A matrix or data.frame.}
}
\description{
Longer proper discription of function...
}
\examples{

## example: ##

# load data
data("snps.ace")
data("phen.ace")
data("tree.ace")
snps <- snps.ace
phen <- phen.ace
tree <- tree.ace

## run ace.test on subset of data (otherwise SLOW!)
out <- ace.test(snps[,1:10], phen, tree, method="discrete")

## examine output
round(out, 4)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

