% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{asr}
\alias{asr}
\title{Ancestral state reconstruction}
\usage{
asr(var, tree, type = c("parsimony", "ML", "ace"),
  method = c("discrete", "continuous"), unique.cols = FALSE)
}
\arguments{
\item{var}{Either a matrix or a vector containing the state of a variable (eg. SNPs or a phenotype)
for all individuals (ie. for all terminal nodes in the tree).}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between the individuals for which snps and phen are known.}

\item{type}{A character string specifying whether ancestral state reconstruction should be
performed by \code{parsimony} or \code{ML} (as performed by the \code{ace} function in package \emph{ape}).}

\item{method}{A character string specifying the type of ASR method to implement,
either \code{'discrete'} or \code{'continuous'} (only used if \code{type} is set to "ML").}

\item{unique.cols}{A logical indicating whether only unique column patterns are present in \code{var}, if \code{var} is a matrix
(if so (TRUE), a time-consuming step can be skipped); by default, FALSE.}
}
\value{
Depending on the dimensions of the input \code{var} object,
either a matrix or a vector containing \emph{both} the known states
of the variable at the terminal nodes (in positions 1:Nterminal) and the
inferred states at internal nodes (in positions (Nterminal+1):Ntotal).
}
\description{
Reconstruct the ancestral states of a vector or matrix object by using either
parsimony or maximum-likelihood methods to infer the states
at the internal nodes of a phylogenetic tree.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
