% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{assoc.test}
\alias{assoc.test}
\title{Run a test of association between SNPs and a phenotype.}
\usage{
assoc.test(snps, phen, tree = NULL, test = c("terminal", "simultaneous",
  "subsequent", "cor", "fisher"), rec = "parsimony")
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{test}{A character string or vector containing one or more of the following available tests of association:
"terminal", "simultaneous", "subsequent", "cor", "fisher". By default, the first three tests are run.
See details for more information on what these tests do and when they may be appropriate.}
}
\description{
Run one of five tests of association between each column of a SNPs matrix and a phenotype
(some tests only implemented for \emph{binary} SNPs and phenotype).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

