% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beeswarm.plot.R
\name{beeswarmPlot}
\alias{beeswarmPlot}
\title{Beeswarm-and-Box-Plot.}
\usage{
beeswarmPlot(y = "sensitivity", x = "test", df, y.lab = NULL,
  pt.size = 4, x.text = FALSE, mean = TRUE)
}
\arguments{
\item{y}{A character string specifying the label of the (numeric) column
in data frame \code{df} to be plotted along the y-axis.}
}
\description{
Wrapper combining the beeswarm and box plot functions from packages \code{beeswarm} and \code{ggplot2}.
}
\details{
NOTE TO SELF--CAREFUL: y.lim is HARD-CODED as (0,1). Change this is you want to release this wrapper!
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

