% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate.integrated_1d_llk_scaled}
\alias{calculate.integrated_1d_llk_scaled}
\title{Function to sum over the LLK grid}
\usage{
calculate.integrated_1d_llk_scaled(prior = prior, llk.surf = llk.surf,
  scaled = TRUE)
}
\arguments{
\item{prior}{Prior}

\item{llk.surf}{log likelihood surface}

\item{scaled}{boolean}
}
\description{
Function to sum over the LLK grid
}
\examples{

}
