% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{calculate_1d_llk_grid_scaled}
\alias{calculate_1d_llk_grid_scaled}
\title{Calculate LLK over the grid}
\usage{
calculate_1d_llk_grid_scaled(prior = NULL, x0 = NULL, x1 = NULL,
  scaled = TRUE)
}
\arguments{
\item{prior}{1D prior}

\item{x0}{Values for unaffected individuals}

\item{x1}{Values for affected individuals}

\item{scaled}{Boolean}
}
\description{
Calculate LLK over the grid
}
\examples{

}
