% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent.tree.sim.R
\name{coalescent.tree.sim}
\alias{coalescent.tree.sim}
\title{Short one-phrase description.}
\usage{
coalescent.tree.sim(n.ind = 100, seed = NULL)
}
\arguments{
\item{n.ind}{An integer specifying the number of terminal nodes desired.}

\item{seed}{An optional integer controlling the pseudo-random process underlying the tree generation.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

