% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_analysis_functions.R
\name{d.llk.d.b0_1d}
\alias{d.llk.d.b0_1d}
\title{Function to calculate d(log likelihood)/db0}
\usage{
d.llk.d.b0_1d(b0 = NULL, b1 = NULL, x0 = NULL, x1 = NULL)
}
\arguments{
\item{b0}{input}

\item{b1}{input}

\item{x0}{input}

\item{x1}{input}
}
\value{
A numeric object
}
\description{
Function to calculate d(log likelihood)/db0
}
\examples{

}
