% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.phen.sim.R
\name{fwd.phen.sim}
\alias{fwd.phen.sim}
\title{Simulate a phenotype, from root to tips.}
\usage{
fwd.phen.sim(snps.assoc, p = 1, tree = NULL)
}
\arguments{
\item{snps.assoc}{A matrix created by the \code{fwd.snp.sim} function,
which indicates where genotypic substitutions occur on the tree at phenoypically-associated sites.}

\item{p}{An integer specifying the probability of phenotypic substition,
given genotypic substitution (see details).}

\item{tree}{An phylo object.}
}
\description{
[*An exploratory function:*] Having already simulated a genotype,
this function allows you to simulate an associated phenotype along the tree, from root to tips.
}
\details{
The parameter \code{p} controls the simulation of the phenotype by specifying
the expected value of the number of phenotypic substitions to occur on the tree provided,
given that a genotypic substitution has occurred on a particular branch of the tree.
}
\examples{

## basic use of fn
tree <- coalescent.tree.sim(n.ind = 100, seed = 1)

## plot output
plot(tree)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
