% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions.R
\name{get.posterior.node}
\alias{get.posterior.node}
\title{Tree Analysis.
Function to check B algorithm and
get credible sets}
\usage{
get.posterior.node(forward, backward, jt.prior, id = 1, plot = FALSE,
  return.ci = TRUE, verbose = FALSE, ci.level = 0.95, log.plot = TRUE)
}
\arguments{
\item{forward}{TODO}

\item{backward}{TODO}

\item{jt.prior}{input}

\item{id}{TODO}

\item{plot}{TODO}

\item{return.ci}{TODO}

\item{verbose}{TODO}

\item{ci.level}{TODO}

\item{log.plot}{TODO}
}
\value{
A data.frame.
}
\description{
Tree Analysis.
Function to check B algorithm and
get credible sets
}
\examples{

}
