% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_analysis_functions_additive_model.R
\name{get.posterior.node_additive}
\alias{get.posterior.node_additive}
\title{get.posterior.node_additive}
\usage{
get.posterior.node_additive(forward, backward, prior, id = 1, plot = FALSE,
  return.ci = TRUE, verbose = FALSE, ci.level = 0.95, log.plot = TRUE)
}
\arguments{
\item{forward}{TODO}

\item{backward}{TODO}

\item{prior}{TODO}

\item{id}{TODO}

\item{plot}{TODO}

\item{return.ci}{TODO}

\item{verbose}{TODO}

\item{ci.level}{TODO}

\item{log.plot}{TODO}
}
\description{
get.posterior.node_additive
}
\examples{

}
