% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{get.sig.snps}
\alias{get.sig.snps}
\title{Get significant SNPs, according to a given test of association.}
\usage{
get.sig.snps(
  corr.dat,
  corr.sim,
  snps.names,
  test = "terminal",
  p.value = 0.01,
  p.value.correct = "bonf",
  p.value.by = "count"
)
}
\arguments{
\item{corr.dat}{A vector containing the association score values, for a given association test, for the real data.}

\item{corr.sim}{A vector containing the association score values, for a given association test, for the simulated data.}

\item{snps.names}{The column names of the original \code{snps} matrix from which the association score values
in \code{corr.dat} were derived.}

\item{test}{A character string or vector containing one or more of the following available tests of association:
"terminal", "simultaneous", "subsequent", "cor", "fisher". By default, the terminal test is run
(note that within treeWAS, the first three tests are run in a loop by default).
See details for more information on what these tests do and when they may be appropriate.}

\item{p.value}{A single number specifying the p.value below which correlations are deemed to be 'significant'.}

\item{p.value.correct}{Specify if/how to correct for multiple testing:
either FALSE, or one of 'bonf' or 'fdr' (indicating, respectively,
the Bonferroni and False Discovery Rate corrections). By default, 'bonf' is selected}

\item{p.value.by}{Specify how to determine the location of the p.value threshold:
either 'count' or 'density' (indicating, respectively, that the p.value threshold should
be determined by exact count or with the use of a density function).}
}
\description{
Identify which SNPs are deemed to be significantly associated with a phenotype,
according to a given test of association and p-value.
(Serves as the treeWAS association testing function;
runs the \code{assoc.test} function internally.)
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
