% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{get.sig.snps}
\alias{get.sig.snps}
\title{Short one-phrase description.}
\usage{
get.sig.snps(snps, snps.sim, phen, test = c("score", "cor", "fisher"),
  p.value = 0.001, p.value.correct = c("bonf", "fdr", FALSE),
  p.value.by = c("count", "density"))
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{snps.sim}{A matrix or list of matrices containing simulated snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{test}{A character string specifying the test to be used to measure correlations between the snps
(real and simulated) and the phenotype. Must be one of 'score', 'cor', 'fisher' (indicating, respectively,
the correlation score, standard correlation, fisher's exact test).}

\item{p.value}{A single number specifying the p.value below which correlations are deemed to be 'significant'.}

\item{p.value.correct}{Specify if/how to correct for multiple testing:
either FALSE, or one of 'bonf' or 'fdr' (indicating, respectively,
the Bonferroni and False Discovery Rate corrections).}

\item{p.value.by}{Specify how to determine the location of the p.value threshold:
either 'count' or 'density' (indicating, respectively, that the p.value threshold should
be determined by exact count or with the use of a density function).}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn

fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

