% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.snps.R
\name{get.sig.snps}
\alias{get.sig.snps}
\title{Get significant SNPs, according to a given test of association.}
\usage{
get.sig.snps(snps, snps.sim, phen, tree, test = c("terminal", "simultaneous",
  "subsequent"), p.value = 0.001, p.value.correct = "fdr",
  p.value.by = "count", snps.reconstruction = snps.REC,
  snps.sim.reconstruction = snps.sim.REC, phen.reconstruction = phen.REC)
}
\arguments{
\item{snps}{A matrix containing the real snps.}

\item{snps.sim}{A matrix or list of matrices containing simulated snps.}

\item{phen}{A factor or vector containing the phenotype (only allowed to contain two levels for now).}

\item{tree}{A phylo object containing a phylogenetic tree in which the number of tips is equal to the
length of \code{phen} and the number of rows of \code{snps} and \code{snps.sim}.}

\item{test}{A character string or vector containing one or more of the following available tests of association:
"terminal", "simultaneous", "subsequent", "cor", "fisher". By default, the first three tests are run.
See details for more information on what these tests do and when they may be appropriate.}

\item{p.value}{A single number specifying the p.value below which correlations are deemed to be 'significant'.}

\item{p.value.correct}{Specify if/how to correct for multiple testing:
either FALSE, or one of 'bonf' or 'fdr' (indicating, respectively,
the Bonferroni and False Discovery Rate corrections). By default, 'fdr' is selected}

\item{p.value.by}{Specify how to determine the location of the p.value threshold:
either 'count' or 'density' (indicating, respectively, that the p.value threshold should
be determined by exact count or with the use of a density function).}
}
\description{
Identify which SNPs are deemed to be significantly associated with a phenotype,
according to a given test of association and p-value.
(Serves as the treeWAS association testing function;
runs the \code{assoc.test} function internally.)
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn

fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

