% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.tip.order}
\alias{get.tip.order}
\title{Get the order of the tip labels of a phylogenetic tree as plotted.}
\usage{
get.tip.order(tree)
}
\arguments{
\item{tree}{An object of class phylo containing a tree
whose tip order is desired to be known.}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
