% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.unique.matrix}
\alias{get.unique.matrix}
\title{Get unique rows/columns of a matrix with an index vector.}
\usage{
get.unique.matrix(data, MARGIN = 2, silent = TRUE)
}
\arguments{
\item{data}{A matrix or data.frame, potentially containing
non-unique patterns in its rows or columns.}

\item{MARGIN}{A single integer specifying the array margin to be held fixed.
(To get unique \emph{rows}, select \code{MARGIN} = 1;
for unique \emph{columns}, select \code{MARGIN} = 2.)}
}
\value{
A list with the following elements:
\itemize{
   \item{\code{index} \item{An index vector containing the indices (row numbers),
         in a matrix composed only of unique rows,
         to which each row in the original matrix maps.}}
   \item{\code{unique.data} \item{A new matrix
         containing only the unique rows of the input matrix.}}
}
}
\description{
A wrapper for the \code{table.matrix} function that assigns consecutive
row or column names to the output matrix's unique rows or columns.
}
\details{
An extension of the base \code{unique.matrix} function,
\code{get.unique.matrix} returns a unique matrix
(by removing duplicate rows or columns), as well as
an index vector that maps each row/column in the original matrix
to the corresponding unique row or column in the deduplicated unique matrix.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
