% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sig.snps.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Short one-phrase description.}
\usage{
manhattan.plot(p.vals, col = "wasp", transp = 0.75, sig.thresh = 0.05,
  thresh.col = "seasun", snps.assoc = NULL, snps.assoc.col = "red",
  jitter.amount = 1e-05, min.p = NULL, log10 = TRUE, ylab = NULL)
}
\arguments{
\item{arg}{Description.}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

