% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phen.R
\name{plot.phen}
\alias{plot.phen}
\title{Short one-phrase description.}
\usage{
plot.phen(tree, phen.nodes, plot = TRUE)
}
\arguments{
\item{tree}{A phylo object.}

\item{phen.nodes}{A vector containing the phenotypic state of either
(i) only terminal nodes in tree or
(ii) all nodes, terminal and internal in tree.}

\item{plot}{A logical specifying whether to display a plot
of the inputted phylogenetic tree with edges coloured to show the
simulated phenotypic substitution process.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

