% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd.plot.prob.phen.R
\name{plot.prob.phen}
\alias{plot.prob.phen}
\title{Short one-phrase description.}
\usage{
plot.prob.phen(p = 0.5, n.snps.assoc = 10)
}
\arguments{
\item{tree}{An phylo object.}

\item{n.subs}{An integer controlling the phenotypic substition rate (see details).}
}
\description{
The parameter n.subs controls the simulation of the phenotype by specifying
the expected value of the number of phenotypic substitions to occur on the tree provided.
The true number of phenotypic substitions is drawn from a Poisson distribution with parameter n.subs.
}
\details{
Longer proper discription of function...
}
\examples{

## basic use of fn ##
## compare probability of having phenotype with 10 SNPs at varying p:
plot.prob.phen(p=0.8, n.snps.assoc=10)
plot.prob.phen(p=0.5, n.snps.assoc=10)
plot.prob.phen(p=0.2, n.snps.assoc=10)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
