% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sig.snps.R
\name{plot.sig.snps}
\alias{plot.sig.snps}
\title{Short one-phrase description.}
\usage{
plot.sig.snps(corr.dat, corr.sim, sig.corrs, sig.snps, thresh, test,
  plot.null.dist = TRUE, plot.dist = FALSE)
}
\arguments{
\item{arg}{Description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn

fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

