% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sig.snps.R
\name{plot_sig_snps}
\alias{plot_sig_snps}
\title{Plot null distribution and significant sites.}
\usage{
plot_sig_snps(corr.dat, corr.sim, corr.sim.subset = NULL,
  sig.corrs = NULL, sig.snps = NULL, sig.thresh = NULL,
  test = NULL, sig.snps.col = "blue", hist.col = rgb(0, 0, 1, 0.5),
  hist.subset.col = rgb(1, 0, 0, 0.5), thresh.col = "seasun",
  snps.assoc = NULL, snps.assoc.col = "red", bg = "lightgray",
  grid = TRUE, freq = FALSE, plot.null.dist = TRUE,
  plot.dist = FALSE, main.title = TRUE, ...)
}
\arguments{
\item{arg}{Description.}
}
\description{
Plot a histogram of the null distribution,
indicating the significance threshold and
the names and association scores of significant sites.
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
