% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFML.R
\name{read.CFML}
\alias{read.CFML}
\title{Convert ClonalFrameML output.}
\usage{
read.CFML(prefix, tree = NULL, plot = TRUE)
}
\arguments{
\item{prefix}{A character string containing the prefix of all file names to be read in.}
}
\value{
read.CFML returns a list containing:
(i) \code{tree}: The phylogenetic tree.
(ii) \code{snps}: The binary genetic data matrix of polymorphic loci.
(iii) \code{snps.rec}: The genetic data reconstruction matrix.
(iv) \code{seqs}: The genetic data sequences (polymorphic loci only), a \code{DNAbin} object.
(v) \code{index}: The index vector, indicating for each column in \code{seqs} the unique polymorphic column pattern to which it corresponds (0 = non-polymorphic).
(vi) \code{n.subs}: The distribution of the number of substitutions per site.
Note that all genetic data elements (ii - iv) are returned in expanded form; that is,
they contain both unique and duplicate column patterns for all polymorphic loci as indicated in the \code{index} vector.
}
\description{
Convert the output of ClonalFrameML into a form usable within \code{treeWAS}.
}
\details{
The \code{prefix} must be the prefix to three files ending in:
(i) "labelled_tree.newick", (ii) "ML_sequence.fasta", (iii) "position_cross_reference.txt".
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
