% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFML.R
\name{read.CFML}
\alias{read.CFML}
\title{Short one-phrase description.}
\usage{
read.CFML(prefix, plot = TRUE)
}
\arguments{
\item{prefix}{A character string containing the prefix of all file names to be read in.}
}
\description{
Longer proper discription of function...
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

