% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous.test.R
\name{simultaneous.test}
\alias{simultaneous.test}
\title{Short one-phrase description.}
\usage{
simultaneous.test(snps.reconstruction, phen.reconstruction, tree)
}
\arguments{
\item{snps.reconstruction}{A matrix containing the terminal and reconstructed
ancestral states of SNPs for all nodes in the tree.}

\item{phen.reconstruction}{A vector containing the terminal and reconstructed
ancestral states of the phenotype for all nodes in the tree.}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between the individuals for which snps and phen are known.}
}
\description{
Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
