% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.sim.R
\name{snp.sim}
\alias{snp.sim}
\title{Short one-phrase description.}
\usage{
snp.sim(n.ind = 100, n.snps = 10000, n.subs = 1, n.snp.assoc = 10,
  assoc.prob = 100, tree = rtree(100), phen.loci = NULL,
  heatmap = FALSE, reconstruct = FALSE, seed = 1)
}
\arguments{
\item{snps}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

