% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent.test.R
\name{subsequent.test}
\alias{subsequent.test}
\title{################
## get.score3 ##
################}
\usage{
subsequent.test(snps.reconstruction, phen.reconstruction, tree)
}
\arguments{
\item{snps.reconstruction}{A matrix containing the terminal and reconstructed
ancestral states of SNPs for all nodes in the tree.}

\item{phen.reconstruction}{A vector containing the terminal and reconstructed
ancestral states of the phenotype for all nodes in the tree.}

\item{tree}{A phylo object containing the tree representing the ancestral relationships
between the individuals for which snps and phen are known.}
}
\description{
########################################################################
}
\details{
###################
## DOCUMENTATION ##
###################

#' Short one-phrase description.
#'
#' Longer proper discription of function...
#'
#' @param Pa A numeric value containing either the state,
#' or the probability of the state, of the phenotype at a given \emph{ancestral} node.
#' @param Pd A numeric value containing either the state,
#' or the probability of the state, of the phenotype at a given \emph{descendant} node.
#' @param Sa A numeric value containing either the state,
#' or the probability of the state, of SNPi at a given \emph{ancestral} node.
#' @param Sd A numeric value containing either the state,
#' or the probability of the state, of SNPi at a given \emph{descendant} node.
#' @param l A numeric value specifying the length of the branch in the phylogenetic tree
#' that joins the ancestral and descendant node.
#'
#' @author Caitlin Collins \email{caitiecollins@gmail.com}
#' @export
#' @examples
#'
#' ## basic use of fn
#' tree <- coalescent.tree.sim(n.ind = 100, seed = 1)
#'

########################################################################

Short one-phrase description.

Longer proper discription of function...
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

