% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent.test.R
\name{subsequent.test}
\alias{subsequent.test}
\title{Short one-phrase description.}
\usage{
subsequent.test(snps.reconstruction, phen.reconstruction, tree)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
tree <- coalescent.tree.sim(n.ind = 100, seed = 1)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

