% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table.matrix}
\alias{table.matrix}
\title{Cross-tabulate the rows or columns of a matrix.}
\usage{
table.matrix(data, MARGIN = 1)
}
\arguments{
\item{data}{A matrix or data.frame, potentially containing
non-unique patterns in its rows or columns.}

\item{MARGIN}{A single integer specifying the array margin to be held fixed.
(To get unique \emph{rows}, select \code{MARGIN} = 1;
for unique \emph{columns}, select \code{MARGIN} = 2.)}
}
\value{
A list with the following elements:
\itemize{
   \item{\code{table} \item{A contingency table of the counts of the
         number of occurrences of each unique row in the matrix.}}
   \item{\code{index} \item{An index vector containing the indices (row numbers),
         in a matrix composed only of unique rows,
         to which each row in the original matrix maps.}}
   \item{\code{unique.data} \item{A new matrix
         containing only the unique rows of the input matrix.}}
}
}
\description{
A version of the base \code{table} function designed for matrices.
Taking a matrix as input, \code{table.matrix} returns a contingency table,
index vector, and unique matrix.
}
\details{
To apply this function to the \emph{columns} of a matrix, simply
transpose the matrix before executing the command, as in:
\code{table.matrix(t(data))}.
}
\examples{
\dontrun{
## load example data:
data("snps.ace")
x <- snps.ace

## basic use of fn on rows of x:
tab.out <- table.matrix(x)

## apply fn to columns of x:
tab.out <- table.matrix(t(x))
}

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
