% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.test.R
\name{terminal.test}
\alias{terminal.test}
\title{Short one-phrase description.}
\usage{
terminal.test(snps, phen)
}
\arguments{
\item{tree}{A phylo object.}
}
\description{
Longer proper discription of function...
}
\examples{

## basic use of fn
out <- terminal.test(snps, phen)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
