% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.reconstruct.R
\name{tree.reconstruct}
\alias{tree.reconstruct}
\title{Short one-phrase description.}
\usage{
tree.reconstruct(dna, method = c("BIONJ", "NJ", "UPGMA", "ML", "BIONJ*",
  "NJ*"), dist.dna.model = "JC69", plot = TRUE)
}
\arguments{
\item{dna}{A binary matrix or DNAbin object containing genomes for (only)
the terminal nodes of the tree to be reconstructed.
Individuals should be in the rows and loci in the columns; rows and columns should be labelled.}

\item{method}{A character string,
one of \code{"NJ"}, \code{"BIONJ"} (the default), \code{"ML"}, or \code{"UPGMA"};
or, if NAs are present in the distance matrix, one of: \code{"NJ*"} or \code{"BIONJ*"},
specifying the method of phylogenetic reconstruction.}

\item{dist.dna.model}{A character string specifying the type of model to use in
calculating the genetic distance between individual genomes (see ?dist.dna).}

\item{plot}{A logical specifying whether to plot the reconstructed phylogenetic tree.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

