% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tree.sim.R
\name{tree.sim}
\alias{tree.sim}
\title{tree.sim
: simulate genotypes based on a given phylogenetic tree.}
\usage{
tree.sim(n.ind = length(tree$tip.label), tree = tree, gen.size = 1000,
  sim.by = "locus", theta = NULL, biallelic = TRUE, seed = NULL,
  plot = TRUE, heatmap = FALSE, plot2 = "UPGMA")
}
\arguments{
\item{n.ind}{The number of individuals in the tree. (Defaults to length(tree$tip.label).)}

\item{tree}{The phylogenetic tree to be used as the basis for the genetic relationship between
the genotypes of the individuals in your simulated dataset.}

\item{sim.by}{The method by which the simulated data is generated (when \code{method = "Didelot"}).
\itemize{
 \item If \code{sim.by = "locus"}, then \code{mt.rate} denotes the expected number of mutations to occur at each locus.
 \item If \code{sim.by = "branch"}, then \code{mt.rate} denotes the expected number of mutations to occur in the whole genome.
 }}

\item{theta}{An optional integer specifying the mutation rate to be used when simulating comparator datasets.
If \code{mt.rate = NULL}, when \code{sim.by = "locus"}, \code{mt.rate} will be set to 1*2;
and when \code{sim.by = "branch"}, \code{mt.rate} will be set to \code{gen.size}*2.}

\item{biallelic}{A logical indicating whether SNPs should be biallelic or otherwise. (TRUE by default.)}

\item{seed}{An optional number to be used as the seed of pseudo-random number generation processes that occur within the \code{tree.sim}
function. Identical pseudo-random processes generated with the same seed will have identical results.}

\item{plot}{A logical indicating whether to plot the phylogenetic tree.}

\item{heatmap}{A logical indicating whether to plot a heatmap of the genetic distances between the individuals in your tree and simulated dataset.}

\item{plot2}{An optional character specifying the method of phylogenetic reconstruction run to confirm that a
 phylogenetic tree which is similar to the inputed phylogeny can be generated based on the simulated genomes alone.}
}
\description{
This function simulates genetic data given an inputed phylogenetic tree structure.
}
\examples{
\dontrun{

# simulate data
out <- coalescent.sim(n.ind=10, gen.size=10000, theta=100*2, biallelic=TRUE, seed=2, plot=TRUE, heatmap=FALSE, plot2="ml")
## isolate elements of out-->
## get list of genomes
x <- snps <- out[[1]]
## get tree
tree <- out[[2]]
## take a look at the phylogeny:
plot(tree)

## run tree.sim
result <- tree.sim(n.ind=length(tree$tip.label), tree=tree,
           gen.size=1000, sim.by="locus", theta=NULL, biallelic=TRUE, seed=NULL, plot=TRUE,
           heatmap=FALSE, plot2="ml")

}
}
\author{
Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}

