% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.R
\name{treeWAS}
\alias{treeWAS}
\title{Short one-phrase description.}
\usage{
treeWAS(snps, phen, n.subs = NULL, tree = c("UPGMA", "nj", "ml"),
  dist.dna.model = "JC69", plot.tree = FALSE, test = c("score", "cor",
  "fisher", "ace.cum", "ace.pagel"), p.value = 0.001,
  p.value.correct = c("bonf", "fdr", FALSE), p.value.by = c("count",
  "density"), sim.n.snps = 10000, n.reps = 1, plot.null.dist = TRUE,
  plot.dist = FALSE)
}
\arguments{
\item{x}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

