% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.R
\name{treeWAS}
\alias{treeWAS}
\title{Short one-phrase description.}
\usage{
treeWAS(x, y, tree = NULL, mt.rate = NULL, p.value = 0.001,
  mt.correct = FALSE, p.value.option = "count", plot.null.dist = TRUE,
  plot.dist = FALSE, plot.Fitch = TRUE, plot.tree = FALSE,
  tree.method = "ml", n.reps = 1, sim.gen.size = 10000,
  sim.by = "locus", dist = NULL, method = "Didelot", test = "score",
  corr.dat.fisher = NULL)
}
\arguments{
\item{x}{description.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

