% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.eval.R
\name{treeWAS}
\alias{treeWAS}
\title{Short one-phrase description.}
\usage{
treeWAS(snps, phen, n.subs = NULL, tree = c("UPGMA", "nj", "ml"),
  dist.dna.model = "JC69", plot.tree = FALSE, test = c("terminal",
  "simultaneous", "subsequent"), p.value = 0.001,
  p.value.correct = c("bonf", "fdr", FALSE), p.value.by = c("count",
  "density"), sim.n.snps = ncol(snps), n.reps = 1, plot.manhattan = TRUE,
  plot.null.dist = TRUE, plot.dist = FALSE, snps.assoc = NULL,
  snps.reconstruction = "parsimony", phen.reconstruction = "parsimony",
  coaltree = TRUE, filename.plot = NULL)
}
\arguments{
\item{test}{A character string or vector containing one or more of the following available tests of association:
"terminal", "simultaneous", "subsequent", "cor", "fisher". By default, the first three tests are run.
See details for more information on what these tests do and when they may be appropriate.}
}
\description{
Longer proper discription of function...
}
\examples{

## load data
data(dist)
str(dist)

## basic use of fn
fn(arg1, arg2)

#' ## more elaborate use of fn
fn(arg1, arg2)

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

