% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeWAS.R
\name{treeWAS}
\alias{treeWAS}
\title{Phylogenetic tree-based GWAS for microbes.}
\usage{
treeWAS(snps, phen, tree = c("BIONJ", "NJ", "UPGMA", "ML", "BIONJ*", "NJ*"),
  n.subs = NULL, n.snps.sim = ncol(snps) * 10, chunk.size = ncol(snps),
  test = c("terminal", "simultaneous", "subsequent"),
  snps.reconstruction = "parsimony", snps.sim.reconstruction = "parsimony",
  phen.reconstruction = "parsimony", na.rm = TRUE, p.value = 0.01,
  p.value.correct = c("bonf", "fdr", FALSE), p.value.by = c("count",
  "density"), dist.dna.model = "JC69", plot.tree = FALSE,
  plot.manhattan = TRUE, plot.null.dist = TRUE, plot.dist = FALSE,
  snps.assoc = NULL, filename.plot = NULL, seed = NULL)
}
\arguments{
\item{snps}{A matrix containing binary genetic data, with individuals in the rows
and genetic loci in the columns and both rows and columns labelled.}

\item{phen}{A vector containing the phenotypic state of each individual, whose length is equal to the number of rows in
\code{snps} and which is named with the same set of labels.
The phenotype can be either binary (character or numeric) or continuous (numeric).}

\item{tree}{A \code{phylo} object containing the phylogenetic tree; or, a character string,
one of \code{"NJ"}, \code{"BIONJ"} (the default), \code{"ML"}, or \code{"UPGMA"},
or, if NAs are present in the distance matrix, one of: \code{"NJ*"} or \code{"BIONJ*"},
specifying the method of phylogenetic reconstruction.}

\item{n.subs}{A numeric vector containing the homoplasy distribution (if known, see details), or NULL (the default).}

\item{n.snps.sim}{An integer specifying the number of loci to be simulated for estimating the null distribution
(by default \code{10*ncol(snps)}). If memory errors arise during the analyis of a large dataset,
it may be necesary to reduce \code{n.snps.sim} from a multiple of 10 to, for example, 5x the number of loci.}

\item{chunk.size}{An integer indicating the number of \code{snps} loci to be analysed at one time. This provides a solution for
machines with insufficient memory to analyse the dataset at hand.
Note that smaller values of \code{chunk.size} will increase the computational time required
(e.g., for \code{chunk.size = ncol(snps)/2}, treeWAS will take twice as long to complete).}

\item{test}{A character string or vector containing one or more of the following available tests of association:
\code{"terminal"}, \code{"simultaneous"}, \code{"subsequent"}, \code{"cor"}, \code{"fisher"}.
By default, the first three tests are run (see details).}

\item{snps.reconstruction}{Either a character string specifying \code{"parsimony"} (the default) or \code{"ML"} (maximum likelihood)
for the ancestral state reconstruction of the genetic dataset,
or a matrix containing this reconstruction if it has been performed elsewhere.}

\item{snps.sim.reconstruction}{A character string specifying \code{"parsimony"} (the default) or \code{"ML"} (maximum likelihood)
for the ancestral state reconstruction of the simulated null genetic dataset.}

\item{phen.reconstruction}{Either a character string specifying \code{"parsimony"} (the default) or \code{"ML"} (maximum likelihood)
for the ancestral state reconstruction of the phenotypic variable,
or a vector containing this reconstruction if it has been performed elsewhere.}

\item{na.rm}{A logical indicating whether columns in \code{snps} containing more than 50\% \code{NA}s
should be removed at the outset (TRUE, the default) or not (FALSE).}

\item{p.value}{A number specifying the base p-value to be set the threshold of significance (by default, \code{0.01}).}

\item{p.value.correct}{A character string, either \code{"bonf"} (the default) or \code{"fdr"},
specifying whether correction for multiple testing
should be performed by Bonferonni correction (recommended) or the False Discovery Rate.}

\item{p.value.by}{A character string specifying how the upper tail of the p-value distribution is to be identified.
Either \code{"count"} (the default, recommended) for a simple count-based approach or
\code{"density"} for a kernel-density based approximation.}

\item{dist.dna.model}{A character string specifying the type of model to use in reconstructing the phylogenetic tree for
calculating the genetic distance between individual genomes,
only used if \code{tree} is a character string (see ?dist.dna).}

\item{plot.tree}{A logical indicating whether to generate a plot of the phylogenetic tree
(\code{TRUE}) or not (\code{FALSE}, the default).}

\item{plot.manhattan}{A logical indicating whether to generate a manhattan plot for each association score
(\code{TRUE}, the default) or not (\code{FALSE}).}

\item{plot.null.dist}{A logical indicating whether to plot the null distribution of association score statistics
(\code{TRUE}, the default) or not (\code{FALSE}).}

\item{plot.dist}{A logical indicating whether to plot the true distribution of association score statistics
(\code{TRUE}) or not (\code{FALSE}, the default).}

\item{snps.assoc}{An optional character string or vector specifying known associated loci to be demarked in
results plots (e.g., from previous studies or if data is simulated); else NULL.}

\item{filename.plot}{An optional character string denoting the file location for
saving any plots produced; else \code{NULL}.}

\item{seed}{An optional integer to control the pseudo-randomisation process and allow
for identical repeat runs of the function; else \code{NULL}.}
}
\value{
A list is returned as the output of \code{treeWAS} containing the data used within it,
and results including information about significant loci identified, if any.
}
\description{
This function implements a phylogenetic approach to genome-wide association studies
(GWAS) designed for use in bacteria and viruses.
The \code{treeWAS} approach allows for the identification of
significant asociations between genotype and phenotype, while accounting for the
confounding effects of clonal population structure,
stratification (overlap between the population structure and phenotypic distribution),
and homologous recombination.
}
\details{
\strong{Data Cleaning}

The genetic data matrix, phenotype, and phylogenetic tree (terminal nodes)
should all be labelled with corresponding sets of names for the individuals.
The order of individuals does not matter, as they will be rearranged to match within the function.

Any individual that is not present in one or more of the genetic data matrix,
the phenotypic variable, and/or the phylogenetic tree must be removed.

If, in the genetic data matrix, redundant columns are present for binary loci
(ie. Denoting the state of the second allele as the inverse of the previous colummn), these should be removed.
For loci with more than two alleles, all columns should be retained. The removal of redundant binary columns
can be done by hand; but, the function \code{get.binary.snps} may also be used. This function requires
column names to have a two-character suffix and unique locus identifiers. The function expects the suffixes
".a", ".c", ".g", ".t" (e.g., "Locus_123243.a", "Locus_123243.g"), though alternative two-character suffixes can be
used (e.g., "Locus_123243_1", "Locus_123243_2") by setting the argument \code{force = TRUE}.
Please also be careful not to accidentally remove any purposeful duplications with repeated names;
for example, if you have deliberately duplicated columns without subsequently generating unique column names
(e.g., by expanding unique columns according to an index returned by ClonalFrameML).

Missing data is permitted (denoted by NA values only) in the genetic data matrix,
but if more than 50% of any column is composed of NAs,
we recommend that this column be removed from the dataset.
Note that the removal of majority-missing columns will be performed automatically within \code{treeWAS}.
If, for some reason, you do not wish this to be the case, set the \code{na.rm} argument to \code{FALSE}.

The phylogenetic tree, if provided by the user, should contain only the terminal nodes corresponding to the individuals under analysis.
Any additional individuals, including the outgroup if not under analysis, should be removed prior to running \code{treeWAS}.
As simulation of the null genetic data is performed along the tree, there is an additional formatting requirement for the tree,
namely that the edge matrix contained in tree$edge must contain, in either the first or last row of its first (ancestral) column,
the first internal node (that is, this node numbered: n.terminal + 1).


\strong{Homoplasy Distribution}

The homoplasy distribution contains the number of substitutions per site.

If this information is not know, it will be reconstructed within \code{treeWAS} using Fitch's parsimony.

If this information is known (i.e., it has been estimated elsewhere through a parsimonious reconstruction),
it can be provided in the \code{n.subs} argument.
It must be in the form of a \emph{named} vector (or table), or a vector in which the \emph{i}'th element
contains the number of \emph{loci} that have been estimated to undergo \emph{i} substitutions on the tree.
The vector must be of length \emph{max n.subs}, and "empty" indices must contain zeros.
For example: the vector \code{n.subs = c(1833, 642, 17, 6, 1, 0, 0, 1)},
could be used to define the homoplasy distribution for a dataset with 2500 loci,
where the maximum number of substitutions to be undergone on the tree
by any locus is 8, and no loci undergo either 6 or 7 substitutions.


\strong{Ancestral State Reconstrution}

If ancestral state reconstruction has been performed
outside of \code{treeWAS} for the \code{snps} and/or \code{phen} variable,
these reconstructions can be submitted in the form of a matrix and a vector, respectively, to the
\code{snps.reconstruction} and \code{phen.reconstruction} arguments.
Please note the formatting requirements.

If provided by the user, \code{snps.reconstruction} should contain \code{snps} in its first \code{nrow(snps)} rows,
and then have the reconstructed states in rows \code{nrow(snps)+1} to \code{nrow(snps)+tree$Nnode}

If provided by the user, \code{phen.reconstruction} should contain \code{phen} in its first \code{length(phen)} elements,
and then have the reconstructed states in elements \code{length(phen)+1} to \code{length(phen)+tree$Nnode}

If created externally, the \code{snps.reconstruction} must have
been generated through either parsimony or ML, and the
\code{snps.sim.reconstruction} argument should be set to match
(as either \code{"parsimony"} or \code{"ML"}), so that a direct comparison can be made.
At least for small datasets, it may be worth (re-)running this
reconstruction within \code{treeWAS} instead, in case any inconsistencies exist between the
external and internal methods of reconstruction.


\strong{Tests of Association}

\emph{Notation used in the equations below:}
           \deqn{G = Genotypic state...}
           \deqn{P = Phenotypic state...}
           \deqn{t = ... at terminal nodes}
           \deqn{a = ... at ancestral nodes}
           \deqn{d = ... at descendant nodes}
           \deqn{Nterm = Number of terminal nodes}

\describe{
\item{\code{terminal}}{The \code{terminal} test solves the following equation, for each genetic locus, at the terminal nodes of the tree only:
                       \deqn{Terminal = | (1/Nterm)*(Pt*Gt - (1 - Pt)*Gt - Pt*(1 - Gt) + (1 - Pt)*(1 - Gt)) |}
                       The \code{terminal} test is a sample-wide test of association that
                       seeks to identify broad patterns of correlation between genetic loci and the phenotype,
                       without relying on inferences drawn from reconstructions of the ancestral states.}

\item{\code{simultaneous}}{The \code{simultaneous} test solves the following equation, for each genetic locus, across each branch in the tree:
                           \deqn{Simultaneous = | (Pa - Pd)*(Ga - Gd) |}
                           This allows for the identification of simultaneous substitutions (i.e., substitutions occuring in
                           both the genetic locus and phenotypic variable on the same branch of the phylogenetic tree,
                           or parallel change on the same branch if non-binary data). Simultaneous substitutions are an
                           indicator of a deterministic relationship between genotype and phenotype. Moreover, because
                           this score measures \emph{only} simultaneous change, and is not negatively impacted by
                           the lack of association on other branches, it may be able to detect associations occurring
                           within some clades but not others and, therefore, to identify
                           loci giving rise to the phenotype through complementary pathways.}

\item{\code{subsequent}}{The \code{subsequent} test solves the following equation, for each genetic locus, across each branch in the tree:
                           \deqn{Subsequent = | 4/3(Pa*Ga) + 2/3(Pa*Gd) + 2/3(Pd*Ga) + 4/3(Pd*Gd) - Pa - Pd - Ga - Gd + 1 |}
                           Calculating this metric across all branches of the tree allows us to measure in what
                           proportion of tree branches we expect the genotype and phenotype to be in the same state.}

   }
}
\examples{

\dontrun{
## load example homoplasy distribution
data(dist_0)
str(dist_0)


## simulate a tree, phenotype, and genetic
## data matrix with 10 associated loci:
dat <- coalescent.sim(n.ind = 100,
                        n.snps = 1000,
                        n.subs = dist_0,
                        n.snps.assoc = 10,
                        assoc.prob = 90,
                        n.phen.subs = 15,
                        phen = NULL,
                        plot = TRUE,
                        heatmap = FALSE,
                        reconstruct = FALSE,
                        dist.dna.model = "JC69",
                        grp.min = 0.25,
                        row.names = NULL,
                        coaltree = TRUE,
                        s = NULL,
                        af = NULL,
                        filename = NULL,
                        set = 1,
                        seed = 1)

## isolate elements of output:
snps <- dat$snps
phen <- dat$phen
snps.assoc <- dat$snps.assoc
tree <- dat$tree

## run treeWAS:
out <- treeWAS(snps = snps,
                phen = phen,
                tree = tree,
                n.subs = dist_0,
                n.snps.sim = ncol(snps)*10,
                test = c("terminal", "simultaneous", "subsequent"),
                snps.reconstruction = "parsimony",
                snps.sim.reconstruction = "parsimony",
                phen.reconstruction = "parsimony",
                p.value = 0.01,
                p.value.correct = "bonf",
                p.value.by = "count",
                dist.dna.model = NULL,
                plot.tree = FALSE,
                plot.manhattan = TRUE,
                plot.null.dist = TRUE,
                plot.dist = FALSE,
                snps.assoc = NULL,
                filename.plot = NULL,
                seed = NULL)

## examine output:
print(out)

}



}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}
